refactored the userRoute to using userController for easier readning
